+++ vitest.config.ts
@@ -4,10 +4,10 @@ export default defineConfig({
   test: {
     include: ['./test/tests/**/*.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
     coverage: {
-      statements: 93,
-      branches: 79.24,
-      functions: 91.66,
-      lines: 93,
+      statements: 91,
+      branches: 78,
+      functions: 87,
+      lines: 91,
       exclude: ['test', 'src/index.ts'],
     },
   },